Rename `object-inspector/utils/` to `/items/`
