+++ src/devtools/packages/devtools-reps/object-inspector/items/container.tsx
