+++ src/devtools/packages/devtools-reps/object-inspector/items/getter.tsx
