+++ src/devtools/packages/devtools-reps/object-inspector/items/loading.tsx
