+++ src/devtools/packages/devtools-reps/object-inspector/items/value.tsx
